adding initial s2i files
this change adds the assemble and run scripts, cleans up the root
directory and readme files, as well as some minor refactoring in the
shiny_squirrel module.

this also restructures the run command to take an environment variable
of MONGODB_SERVICE_NAME instead of expecting it as a passed in argument.